refactor: rename exception classes
