+++ src/exceptions/DatabaseException.ts
@@ -1,4 +1,4 @@
-import HttpException from "./Http";
+import HttpException from "./HttpException";
 
 export default class DatabaseException extends HttpException {
     status: number;
