+++ src/middleware/ErrorHandler.ts
@@ -1,5 +1,5 @@
 import { NextFunction, Request, Response } from 'express';
-import HttpException from '../exceptions/Http';
+import HttpException from '../exceptions/HttpException';
 
 const errorHandler = (error: HttpException, _request: Request, response: Response, _next: NextFunction) => {
     const status = error.status || 500;
